Added rbac policy to allow running privlieged containers, Improved readability
